SMB: Modify SMB stateless class to utility functions
